requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
Faker==18.13.0
fastapi==0.100.0
FastAPI-SQLAlchemy==0.2.1
fastapi-users==12.0.0
fastapi-users-db-sqlalchemy==5.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
jsf==0.7.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
minio==7.1.15
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
rstr==3.2.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlalchemy2-stubs==0.0.2a34
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
